Refactor entities and update database schema
- Removed `using` directives and reformatted namespaces in multiple DTO and command files.
- Changed `IMovieListRepository` interface to use `ShareList` instead of `MovieList`.
- Removed `IUserEntityRepository.cs`, `UserEntity.cs`, and `MovieList.cs` files.
- Updated `Recents.cs` to use `ApplicationUser` and removed `UserEntityID`.
- Added new package reference for `Auth` in `Core.Domain.csproj`.
- Updated `KhakuContext.cs` to replace `MovieList` with `ShareList` and `UserEntity` with `ApplicationUser`.
- Updated `KhakuContextModelSnapshot.cs` to reflect model changes.
- Updated `MovieListRepository` to use `ShareList`.
- Removed `UserEntityRepository` and its registration in `ServiceRegistration.cs`.
- Added new `ShareList` class in `Core.Domain.Entities.UserThings`.
- Added new migration file `20250120013125_User logic changed.cs` for schema changes.
- Added `ShareListController` with `GetHome` endpoint.
