+++ Core.Application/DTOs/Genres/TmdbGenreResponseListDto.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Core.Application.DTOs.Genres
+﻿namespace Core.Application.DTOs.Genres
 {
     public class TmdbGenreResponseListDto
     {
