+++ Core.Application/DTOs/Movies/MovieSearchModuleDto.cs
@@ -1,9 +1,4 @@
 ﻿using Core.Application.DTOs.Genres;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Core.Application.DTOs.Movies
 {
