+++ Core.Application/Features/GenreModule/Commands/GetGenresFromAMovie/GetGenresFromAMovieCommand.cs
@@ -1,13 +1,7 @@
 ﻿using AutoMapper;
-using Core.Application.DTOs.General;
 using Core.Application.DTOs.Genres;
 using Core.Application.Interface.Repositories;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Core.Application.Features.GenreModule.Commands.GetGenresFromAMovie
 {
