--- Core.Application/Interface/Repositories/IUserEntityRepository.cs
@@ -1,8 +1,0 @@
-﻿using Core.Domain.Entities.User;
-
-namespace Core.Application.Interface.Repositories
-{
-    public interface IUserEntityRepository : IGenericRepository<UserEntity>
-    {
-    }
-}
