+++ Core.Domain/Core.Domain.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Auth" Version="0.0.0-alpha.0.29" />
+  </ItemGroup>
+
 </Project>
