+++ Core.Domain/Entities/UserThings/Recents.cs
@@ -1,16 +1,16 @@
-﻿using Core.Domain.Common;
+﻿using Auth.Infraestructure.Identity.Entities;
+using Core.Domain.Common;
 using Core.Domain.Entities.GeneralMovie;
-using Core.Domain.Entities.User;
 
 namespace Core.Domain.Entities.UserThings
 {
     public class Recents : AuditableBase
     {
         public DateTime Date { get; set; }
-        public int UserEntityID { get; set; }
+        public string UserEntityID { get; set; }
         public int MovieID { get; set; }
 
         public Movie Movie { get; set; }
-        public UserEntity UserEntity { get; set; }
+        public ApplicationUser User { get; set; }
     }
 }
