+++ Infraestructure.Persistence/Repositories/MovieListRepository.cs
@@ -4,7 +4,7 @@ using Infraestructure.Persistence.Context;
 
 namespace Infraestructure.Persistence.Repositories
 {
-    public class MovieListRepository : GenericRepository<MovieList>, IMovieListRepository
+    public class MovieListRepository : GenericRepository<ShareList>, IMovieListRepository
     {
         private readonly KhakuContext _dbContext;
