+++ Infraestructure.Persistence/ServiceRegistration.cs
@@ -27,7 +27,6 @@ namespace Infraestructure.Persistence
             services.AddTransient<IMovieWebRepository, MovieWebRepository>();
             services.AddTransient<IRecentsRepository, RecentsRepository>();
             services.AddTransient<IScrapPageRepository, ScrapPageRepository>();
-            services.AddTransient<IUserEntityRepository, UserEntityRepository>();
         }
     }
 }
