+++ Presentation.KuhakuCentral/Controllers/V1/ShareListModule/ShareListController.cs
@@ -4,18 +4,18 @@ using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
 using System.Net.Mime;
 
-namespace KuhakuCentral.Controllers.V1.HomeModule
+namespace KuhakuCentral.Controllers.V1.ShareListModule
 {
-    public class HomeModuleController : BaseAPI
+    public class ShareListController : BaseAPI
     {
         [HttpGet("GetHome")]
         [Consumes(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [SwaggerOperation(
-            Summary = "Home Page",
-            Description = "Get Home Page Data"
-            )]
+                    Summary = "Home Page",
+                    Description = "Get Home Page Data"
+                    )]
         public async Task<IActionResult> Home(bool KidMode = false)
         {
             return Ok(await Mediator.Send(new GetHomePageDataQuery() { KidMode = KidMode }));
